Issue-10 create internal package structure and move database functionality
